rename factory function files
