+++ src/renderer/StratifiedSampler.js
