+++ src/renderer/TextureAllocator.js
